Rework parsing according to the new specs, add parameterization
This adds support for a new parameterization syntax that is highly
similar to the way Om Next does it. Parameterization is possible
by writing

    (<property or nested properties> <map of params to values>)

Example queries are:

    [(user [name email] {user-id ?id})]

    [([user-info _] {user-id ?id})]

    [({users [db [id] user [name email]]}
      {name-matches? ?search-str})]

In these examples, ?id and ?search-str are variables that can be
bound by whatever consumes the query definition. In the case of
Om Next components defined with defview, ?id would be bound to a
component query param.

The values of parameters can be anything, not just variables.