Rename parse to conform-and-parse, add separate parse function
We need this separation because sometimes we want to parse raw
queries that first need to be conformed and other times we already
have conformed queries and want to parse them directly.

Make defview use the new conform-and-parse function instead of
parse.

Use the new conform-and-parse function in tests instead of parse.